+++ src/main/java/dev/jpaboard/user/api/UserController.java
@@ -1,28 +1,19 @@
 package dev.jpaboard.user.api;
 
-import dev.jpaboard.common.exception.AuthorizedException;
 import dev.jpaboard.user.application.UserService;
-import dev.jpaboard.user.dto.request.UserCreateRequest;
-import dev.jpaboard.user.dto.request.UserLoginRequest;
 import dev.jpaboard.user.dto.request.UserUpdateRequest;
 import dev.jpaboard.user.dto.response.UserResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.SessionAttribute;
 
-import java.util.Objects;
-
 import static org.springframework.http.HttpStatus.NO_CONTENT;
-import static org.springframework.http.HttpStatus.OK;
 
 @RestController
 @RequestMapping("/api/users")
@@ -31,35 +22,7 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping("/sign-up")
-    @ResponseStatus(OK)
-    public void signUp(@RequestBody UserCreateRequest request) {
-        userService.signUp(request);
-    }
-
-    @PostMapping("/login")
-    @ResponseStatus(OK)
-    public Long login(@RequestBody UserLoginRequest request, HttpServletRequest httpRequest) {
-        httpRequest.getSession().invalidate();
-        HttpSession session = httpRequest.getSession(true);
-
-        Long userId = userService.login(request);
-        session.setAttribute("userId", userId);
-        session.setMaxInactiveInterval(3600);
-        return userId;
-    }
-
-    @PostMapping("/logout")
-    @ResponseStatus(NO_CONTENT)
-    public void logout(HttpServletRequest httpRequest) {
-        HttpSession session = httpRequest.getSession(false);
-        if (Objects.isNull(session)) {
-            throw new AuthorizedException();
-        }
-        session.invalidate();
-    }
-
-    @PatchMapping("/update")
+    @PatchMapping
     @ResponseStatus(NO_CONTENT)
     public void update(@RequestBody UserUpdateRequest request,
                        @SessionAttribute(name = "userId") Long userId) {
