+++ src/main/java/dev/jpaboard/user/application/UserService.java
@@ -1,8 +1,6 @@
 package dev.jpaboard.user.application;
 
 import dev.jpaboard.user.domain.User;
-import dev.jpaboard.user.dto.request.UserCreateRequest;
-import dev.jpaboard.user.dto.request.UserLoginRequest;
 import dev.jpaboard.user.dto.request.UserUpdateRequest;
 import dev.jpaboard.user.dto.response.UserResponse;
 import dev.jpaboard.user.exception.UserNotFoundException;
@@ -17,18 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-
-    @Transactional
-    public UserResponse signUp(UserCreateRequest request) {
-        User user = UserCreateRequest.toUser(request);
-        User savedUser = userRepository.save(user);
-        return UserResponse.from(savedUser);
-    }
-
-    public Long login(UserLoginRequest request) {
-        User user = findByEmailAndPassword(request.email(), request.password());
-        return user.getId();
-    }
 
     @Transactional
     public UserResponse update(UserUpdateRequest request, Long userId) {
@@ -49,11 +35,6 @@ public class UserService {
 
     private User findUserById(Long userId) {
         return userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
-    }
-
-    private User findByEmailAndPassword(String email, String password) {
-        return userRepository.findByEmailAndPassword(email, password)
                 .orElseThrow(UserNotFoundException::new);
     }
