+++ src/main/java/dev/jpaboard/user/dto/request/UserCreateRequest.java
@@ -4,13 +4,13 @@ import dev.jpaboard.user.domain.User;
 
 public record UserCreateRequest(String email, String password, String name, int age, String hobby) {
 
-    public static User toUser(UserCreateRequest request) {
+    public User toUser() {
         return User.builder()
-                .email(request.email)
-                .password(request.password)
-                .name(request.name)
-                .age(request.age)
-                .hobby(request.hobby)
+                .email(this.email)
+                .password(this.password)
+                .name(this.name)
+                .age(this.age)
+                .hobby(this.hobby)
                 .build();
     }
